Used int16_t throughout scroll module.

+++ scroll.h
@@ -4,8 +4,8 @@
 #define JBXVT_SCROLL_H
 #include <xcb/xcb.h>
 struct JBXVTLine * jbxvt_get_saved_lines(void);
-int jbxvt_get_scroll_size(void);
+int16_t jbxvt_get_scroll_size(void);
 void jbxvt_clear_scroll_history(void);
-void scroll(xcb_connection_t * xc, const int row1,
-	const int row2, const int count);
+void scroll(xcb_connection_t * xc, const int16_t row1,
+	const int16_t row2, const int16_t count);
 #endif//!JBXVT_SCROLL_H
